+++ desc_tokenize.rb
@@ -1,96 +1,27 @@
 require './mnt'
-require 'uri'
-class DescriptionParser
-def parse_tokens(string)
-  # split after urls 
-  regexp = /[🔥🌿🏓;♥️✪♡／|•☾✚★∆✗\{\}\(\)\]\[]|[.!? ] | [\/ :]{2,} | [~*·\-\/,] |\r\n|\n/
-  string.split(regexp).
-  map(&:strip).
-  select{ |v|
-    v && v.length > 0
-  }
-
-end
-
-def word_split(string)
-  # break after seperator is 'and' or '&'
-  split_regexp = /[;,&]|\Wand\W/
-  string_regexp = /.\w\W(on|to|by|is|of)\s/
-  if string =~ string_regexp 
-    i = string.index string_regexp 
-    {
-      string[0..(i+4)] =>
-      string[(i+5)..-1].split(split_regexp).map(&:strip).map{ |s|
-        s ? word_split(s) : s
-      }
-    }
-  else
-    { string => nil }
-  end
-end
-
-def words_split(string)
-  result = {} 
-  n= string.map{ |e|
-    result = result.merge word_split(e)
-  }
-  result
-end
-
-def colon_split(string)
-  split_regexp = /[;,&]|\Wand\W/
-  colon_regexp = /\w: | =/
-  if string =~ colon_regexp
-    i = string.index(colon_regexp)
-    {
-      string[0..(i)] =>
-      string[(i+2)..-1].split(split_regexp).map(&:strip).flatten
-    }
-   else
-     { string => nil}
-   end
-
-end
-
-def colons_split(string)
-  result = {}
-  n= string.map{ |e|
-    result = result.merge colon_split(e)
-  }
-  result
-end
-
-def string_to_mappings(strings)
-  m = colons_split(strings) || {}
-  n = words_split(strings) || {}
-  result = n.merge m
-  puts '#' *50
-  puts 'strings', strings
-  puts 'result.inspect',result.inspect
-  result
-end
-
-def tokens_to_mappings(tokens)
-# implement lists in language, list:
-# one, two and three
-mts = tokens.map{ |k,t|
-  [k,string_to_mappings(t)]
-}.to_h
-mts
-end
-end
+require './parser'
+p = Parser.new
 descriptions = {}
-DB.from(:user).group_by(:description).order_by(:db_id).limit(100000).where("description IS NOT NULL").map{|d| descriptions[d[:screen_name]] = "#{d[:description]} \r\n #{ descriptions[d[:screen_name]]}" }
+per = 100
+100.times do |i|
+  DB.from(:user).group_by(:description).order_by(:db_id).limit(per,i*per).where("description IS NOT NULL").map{|d| descriptions[d[:screen_name]] = "#{d[:description]} \r\n #{ descriptions[d[:screen_name]]}" }
+end
 #tokens = descriptions.select{|k,d| d }.map{ |k,d| t=d.split(/[♥️✪|•☾✚∆✗]|[.!?] | \/\/ | [-\/] |\r\n|\n/).map(&:strip); puts t.inspect ;[k,t] }
 #tokens = descriptions.select{|k,d| d }.map{ |k,d| t=d.split(/[♥️✪|•☾✚∆✗]|[.!?] | \/\/ | [-\/,] |\r\n|\n/).map(&:strip); puts t.inspect ;[k,t] }
 #tokens = descriptions.select{|k,d| d }.map{ |k,d| t=d.split(/[♥️✪|•☾✚∆✗]|[.!?] | [\/ ]{2} | [-\/,] |\r\n|\n/).map(&:strip); puts t.inspect ;[k,t] }
-p= DescriptionParser.new
 tokens = descriptions.select{|k,d| d }.map{ |k,d|
   [k,p.parse_tokens(d)] 
 }
 
 tokens = tokens.to_h
 mts = p.tokens_to_mappings(tokens)
+
+# puts 'mts.inspect', mts.inspect
+#tokens.each{ |k,v| 
+#  puts k.rjust(30,'~'), v.map{ |m,n| 
+#    "#{m.ljust(40)} #{n}"
+#  } 
+#}
 
 #mts.each{ |k,v| puts k, v.map{|m,n| "#{m}: \t #{n}"} };nil
 #mts.each{ |k,v| puts k.rjust(30,'~'), v.map{|m,n| "#{m.ljust(40)} #{n}"} };nil
